part 1 getOrder

+++ mocktest3/mocktest3/Models/DTOs/OrderDto.cs
@@ -4,7 +4,7 @@ public class OrderDto
 {
     public int Id { get; set; }
     public DateTime Date { get; set; }
-    public double TotalPrice { get; set; }
+    public decimal TotalPrice { get; set; }
     public List<ProductDto> Products { get; set; }
 }
 
@@ -12,5 +12,5 @@ public class ProductDto
 {
     public string Name { get; set; } = string.Empty;
     public int Amount { get; set; }
-    public double Price { get; set; }
+    public decimal Price { get; set; }
 }
